kafka: changed kafka_config to work with remote tunneling
What happens is kafka on the remote only knows itself by kafka.tls.ai,
its a similar problem to when trying to connect via ssl to a remote. So
the kafka config has to have kafka.tls.ai and then for it to go out via
the tunnel, kafka.tls.ai has to be configured to localhost and then
traffic to kafka.tls.ai:local_bind_port goes out via the tunnel to the
remove kafka client.