kafka: bugfix for communicating with kafka tunneled to remote
The plugin didnt used to work when communicating tunneled to a remote
machine. This was because kafka always communicates to kafka.tls.ai on
port 9092 and the tunnel was going to a random local_bind_port, so I
massaged the infra to allow manually setting the local_bind_port and now
tunneling works.

In addition, added kafka.tls.ai 127.0.0.1 to local
container hosts file so the whole thing works. Previously this worked by
having an add_host param to docker run command, but we didnt want to
have to add this for other plugins, rather decided hosts file should be
edited and add itself to hosts file (inside the docker container).

1 last thing, added a basic test to verify the flow works by creating a
temp topic, writing to it, comsuming it, and then deleting it.