K8s plugin | cli wrapper for k8s
This is a plugin which warps the k8s cli and executes it's command via
SshDirect plugin.
This plugin uses a util function called command_builder which basically
just allows use to join all the args passed to function and format them
for execution.

List of functions:
1. create: Creates any k8s resources the caller wants
3. scale, expose, get, apply: do their corresponding command in the k8s
cli
4. count pods: counts the amount of pods current running in
k8s we use this function to verify the state of the k8s before and after
the test and that action like creating pods successed
5. get version: gets the version the k8s running which is required since
the pod ip of each pod is saved in a different place depending on
version of k8s
6. get ready deployment pods: Check if all pods of a certain deployment
are in a ready status
7. pod is alive: curl's to an ip and returns the response and in case
the response is an ssh error it prints it
8. get pods internal ip: Attempts to get the the internal ip of the pod
and since the location of the internal ip within the pod object has
changed from k8s version 1.14.8 and 1.17.0 it search in a different
location within the pod object depending on the k8s version
9. clean up deployment: attemps to deletes a deployment and it's
corresponding svc

K8s plugin | Support kwargs for higher level functionv

Now the following functions can be passed kwargs and run them ask
options to the underlying k8s cli command.
- delete_deployment
- delete_svc
- get_deployment
- create_deployment
- scale_deployment
- expose_deployment

Example: delete_svc(name,now=False )
would execute
sudo gravity exec kubectl delete name --now=False

Bug Fix | k8s version between 1.14.8 and 1.17.0 would fail the test

in case the k8s version was between 1.14.8 and 1.17.0 the test would
fail to get pod if's since it wouldn't match any of the expected if
statements

k8s plugin create | added wait to deployment creation