Test_k8s_networking | Verify flannel bug scenario
This test Deploys a dummy app replicates it to a give amount of
replicas, then Reboots the master node of the k8s cluster master.
gets all the internal ips of the dummy deployment pods and attempts to
make a simple curl ip:port/ expecting each pod will return the expected
respone and if not fails the test indicating that one of the pods wasn't
available.

The test assumes the host under test is the master of the k8s cluster

The test doesn't clean up before or afte since that will be the job of a
higher level cleaner which will run before each test of the devops infra

The k8s utils simplifies the creation of deployment with given amount of
pods and expose all of it's pods so they host under test can reach them

The is cluster healthy gravity function checks if the whole cluster is
healthy and only passes if it's

The number of health nodes function function returns the amount of nodes
which are healthy which will be usefull for future test