rpyc-kafka-server: added create/delete/get topics functions
Despite this functionality being available to the client via the
kafka.admin property, for some reason its super slow when calling
locally (~40s to create or delete 25 topics).
Adding remote functionality lowers the create/delete duration ~0.5-3s,
so a drastic improvement.
Because we need to delete the topics between each tests to clean the
kafka, we need this functionality to be super fast.