promethous: fix plugin connectivity in k8s
when trying to work with promethous plugin on w8s environment, getting resolving issue.
this cause because the containerize is not familiar with prom.anv dns name.
the soultion was to write into hosts file the:
<host ip> prom.anv
why we can't work directly to prom in 9090 in k8s? because prom is inside the cluster and exposed only by ingress.
if we wanted to use the automation proxy container we had to develop another rpyc server. but for this case this is enough

issue: https://anyvision.atlassian.net/browse/TMTNNFR-172