Util | fake pois
In order to test a cluster under normal load we require the ability to
insert a large amount of pois for the reid to work on this code is added
in order to create 10k+ pois on demand (Insertion code written by yoni
the legend).

Becuase of current speed issues with executing the query using the
plugins we are executing the insertion query by just running a script
which generates the fake pois and inserts them into the DB on the
automation pod it self.