cleaners: add option to explicitly clean storage in composie
I`m adding methods that will clean storage in its primitive way, this is
usefull as implementing "smart" cleaners that clean specific parts of
"data services" isnt always enought.
For example for consul we need to delete all services registration, kv
and maybe other stuff. Also in kafka case it was easier to actually to
clean the storage