consul: add api to get values casted to type.
There is a new api added to consul tht will allow to get/set
float/ints/jsons directly without a need from application to cast it
json.loads("2.5") is 2.5, so every plain type we can just load from json
and it will be casted. It does not work with strings i.e. json.loads("bla")
yeilds an exception so we have to deal with string type separately