seaweed: inherit from resource_manager instead of the opposite.
Up until now, the resource_manager plugin inherited from the seaweed
plugin (which sits in devops-infra repo), in order to take advantage of
some methods which were implemented there, and because the resource
manager is actually also an s3 object storage.
The problem with that is that it creates a dependency for infra repo on
devops-infra repo, when really it should be the other way around, ie
devops-infra (and seaweed) can depend on infra. So this would cause
test_resource_manager to fail in a situation where it wasnt cloned
together with devops-infra repo.

The solution was to move the shared methods to the resource_manager
plugin, and change the seaweed plugin (in devops-repo) to inherit from
resource_manager.