infra: fix copy files from host running tests to container.
Currently writing to container was incorrect and worked only in case
host running test and host under test were running on the same machine.
The reason for that was that we cannot copy file from remote host to the
docker running on some other host. File must be first copied to the
"host under test" and then moved to the container itself.
Fixes: a1757a9 ("add copy file to container function")