Docker plugin : add functionllity to wait for status x times
currently we have problem with restart container
wait_for_container up not giving in this situation required behavior.
cause container might continue to reset himself and we won't catch it in
wait_for_container_up.
in this new function we check container is up x times with docker
inspect status