tunnel: fix the ip of the tunnel we create.
Currently the endpoint of the tunnel is defined as the host under test,
this is incorrect as the endpoint of the tunnel is actually "local" to
that host (it is automation infra container that is running kafka proxy
server).
This does not work on AWS, where the external IP of the host cannot be
used for host to host communication. (external IP is not resolvable from
within the host itself), and therefore tunnel target cannot be the
external ip of the host