weed: solve OOM issue when passing large artifacts from AWS to local S3.
Currently we have an OOM in our tests  when passing large artifacts from
AWS to local S3. this is caused due to the fact that we first read the
whole object to memory and then try to upload it to local s3. For large
files this obviously does not work and we get an OOM.
Solution is to use multipart download/upload to avoid keeping the object
as a whole in memory