kafka: fix tunnel endpoint of kafka rpyc proxy.
Kafka plugin works in the following way:
1) We have a python rpyc server running on automation proxy container
(host under test)
2) Python client that runs on host running test connects to that python
code (rpyc kafka) via ssh runnel (which is also running on automation
proxy).
3) the target of the tunnel was set to localhost, which means that
automation_proxy was trying to open a connection to "localhost". This is
correct, but!!

We had a corrputed /etc/hosts which did not contain an entry for
localhost, which caused localhost not been resolved from
automation_proxy container.
This caused tunnel not been able to be established to the "kafka python
rpyc" server and all kafka related tests to fail on such setups.

Now to be safe we just change it to 127.0.0.1 to avoid DNS issues
enterely.

Fixes: 1f2f1f8 ("tunnel: fix the ip of the tunnel we create.")