renamed devops_product_manager to devops_docker_installer and updated
The name is more fitting, that way we know its the installer for docker,
and in the future when we will have a k8s installer we will be able to
differentiate.
In addition, updated it to suit v2_infra, so base the installation on
pytest_after_proxy_container hook, instead of implementing as a session
scoped fixture.