devops_proxy_container pytest plugin added
Previously the automation_proxy was rolled up inside the
pytest_autoamtion_infra plugin (in autaomtaion-infra) and would run no
matter what HUT we were using.
This is a bad idea because there could be tests in which we dont want
the proxy container to run, or for example on an edge device where the
is no docker.

So we decoupled the proxy container from automation-infra, and put the
logic in the devops infra.

This is the pytest_plugin which needs to be invoked to get the
automation_proxy container to run on the HUT.