docker registry: add docker registry plugin.
What is this used for:
When developing service from machine that is different from the
machine the test is running it is quite tricky to test it.
In order to do that one needs to submit it to some central registry
(upload which is slow), then somehow pull it on target machine, re-run
the whole installation with new modified image.
This is quite cumbursome development cycle.
The solution: Upload image from host running test to host under test in
runtime.
The problem: There is no docker registry on host running test.
The solution: We will deploy one. Now how do we do it ? as registry is
secured and we wont just be able to push images to it. Well the
workaround is to create a tunnel from host running test to host under
test, then submit the image to "localhost" which is actually tunelled to
host running test. This way we bypass the security of the docker :)