devops_proxy_container: clean_between_tests tryfirst
Its important for devops cleaner to run before any other cleaners bc it
reset proxy container, so any tunnelled connections will be lost.

For example if core cleaner runs and starts a kafka rpyc server and
pings to it, but then afterwards the container is restarted, the kafka
rpyc server will no longer run, but the plugin will have the rpyc saved,
and therefore will no longer work.

The solution is to tryfirst the devops cleaner, and all the rest can
come afterwards.