run/bash scripts: add --noconftest flag
With v3 infra was can use conftest to add installers or implementations
of gossip hooks, which we dont want to be triggered on main (parent)
pytest call, rather only on secondary pytest call.

Adding --noconftest flag to the main call will facilitate this.
We already were doing this on core-product run scripts, but this wasnt
necessary before as only now did we add conftest files to devops test
directories.