devops_docker/conftest.py: move pytest_addoption hook to conftest.py
This needs to be moved and cant be in both places otherwise pytest will
raise an error of "option names already added" since it is being added
in 2 places.

In theory we would have liked to implement the add_options via gossip
hook, triggered in pytest_automation_infra. Unfortunately this is not possible
because pytest starts a new process between the addoptions and the
actual pytest run.

Starting a new process is problematic because of the way gossip is implemented,
that when a module containing gossip hook implementations is imported
they are registered, but then when a new process is started, the
registrations are obviously cleared.

At this stage the only solution is to add_options via pytest hook, and
all other functionality can be added via gossip hooks.