installers/ssh.py: register to setup stage
Because clean_between_tests disconnects SshDirect plugin, it is
necessary to reconnect in the setup stage.

Previously this worked because the docker.py installer did it, and it
still will need to, but tests which dont invoke the "docker" installer
(we dont have many right now) would have failed. Now they will pass.