proxy_container.py: added clear() method
in automation-infra repo we added functionality to host.clear_plugins
function which calls the plugin.clear for each plugins before it delets
the plugin itself. This is to enable a plugin to clean up leftovers if
it wants to. So for the proxy_container.. would be good to kill it so
its not leftover.. if someone else needs it - deploy it.