memsql: fix memsql active state check.
Currently we checked memsql active state by running show dbs command.
However we got a failure:
'Partition tracks_db:0 has no master instance. This is likely because the node or nodes that hold a copy of the partition are down. Check for offline leaf nodes by running SHOW LEAVES and bring them back online to restore access to the partition')

This is probably happening due to the incorrect check of "memsql" up.
Changed this to check the leaves state