Kafka plugin: Disable strimzi cluster operator before deleting kafka's data
When scaling down Kafka's StatefulSet in order to delete it's pvc,
he operator enforces the original replicas configuration in the CRD
causing the scale operation to fail and get a timeout.

To prevent this behaviour, strimzi cluster operator needs to be disabled
before scaling down kafka.

Changing the replicas in the CRD didn't work for some reason.
For now, we'll use this workaround