K3s installer: ignore initial k3s error
In order to start k3s server we need to restore etcd data from sanpshot.
The initial start command results with error because the service isn't
starting, it needs to have another reset without 'cluster-reset' and
'cluster-init' flags.

The error caused the installation to fail, this commit ignores the error
then continues with the second restart which is needed to start the
server