Kubectl utils: Fixed timing issue when deleting StatefulSet data
Sometimes, the scaling up the statefulSet right after deleting it's pvc
causes the pod to stuck on pending state because it can't find it's pvc,
although it should create the pvc by itself.

To make sure it doesn't happen, added a waiter function which waits
until the pvc is not exist before scaling up the statefulSet.