:bug: K3s Setup | Ensure rancher pod is setup with renewed certificate
Rancher certificate is provisioned with Cert manager,
which means it rolls out every 3 months.
This caused an issue when initiating rancher-cli:

Rancher login was running before the certificate has been rolled out therefore,
The login failed due to certificated expired error.

Added a logic that checks if the certificate is expired and if so, it waits for it
to be renewed, and then restarts Rancher pod. This way the cluster setup ensures that
the rancher pod is setup with the new certificate before login.